A lot of enhancements implemented while being offline:
- Adapted BuildRoot and CompressAndBuildISO to update packages (and other items) on ROLayer, avoiding the need to remove the whole ROLayer directory
every time an external tarball is changed (md5sum is used). Warning: assuming there is only one version per program (can be improved in the future,
if needed).

- Removed the "External" directory, which contained unpacked tarballs. Now, CompressAndBuildISO is getting the files always directly from the
tarballs when inserting them in ROLayer and initrd. This avoids unnecessary local replication, removes a possible point of failure and allows a
better control of what has and what has not to be updated on an existing ROLayer. As a side effect, "mode" does not need to be set at
'Packages-Devel-i686', since only 'fetch' mode is used now.

- Cleanup on ChrootEnvironment: removed most of the 'convenience binds', since they can be harmful (e.g., rm -rf follows the binded dirs) and they
are not so useful anymore (most of the work is now done without having to be inside ChrootEnviroment). Only Chroot_Mount_CD-ROM bind was kept.

- Removed Master.conf: the directories listed at Master.conf were already hardcoded within the scripts. Don't having to deal with this configuration
file simplifies the overall idea of BuildLiveCD scripts. We can re-add Master.conf latter, with a better support on the scripts. Btw, a concept
similar to layering is already implemented with separated squashfs files and packages lists (stored at Data directory).

- Moved RWLayer to Output, since having it at the same level as ROLayer makes more sense (IMHO).

- Creating most directories "on-demand" (tm). E.g.: Output/RWLayer is now created only on ChrootEnvironment; Output/Enviromnent is created and
removed at each ChrootEnvironment call.

- Better isolation of each phase on both UpdateEnvironment and CompressAndBuildISO.

- UpdateEnvironment has now only download-related tasks. It gets Archives, CVS and Packages. (Unpacks and directory creations - other than
Archives/* and LocalCVS - are not needed anymore.)

- Removing old gobo-tools snapshots each time a new version is packed.

- Updated CD-Bootloader-i686.tar.bz2 file, pushing files one directory up (removed root i686 directory: isolinux and README are now at top level);
(Having README file inside bootloader tarball does not make much sense, but it is easy to change this latter.)