Removed --repack option.
Changed the way unpacked packages are handled: instead of unpacking them directly inside Clean or Bootstrap, they are always unpacked to Unpacked.
Control of when a package inside Unpacked needs to be reunpacked is done with md5sum, like it is done in CompressAndBuildISO.

Tools (Compile and Scripts) snapshots are now packed and unpacked, instead of directly copied from cvs.