ContributeRecipe: code cleanup and minor fixes
Use the '$@' syntax so we don't have to escape arguments provided to
the new functions and shorten the retry loop implementation. The change
to the '$@' syntax as opposed to a call to 'bash -c' also prevents a bug
in which the call to 'cd $recipeDir' has no effect (because it runs on
a separate shell).