Fixes unexpected behavior of Compile with overlayfs
SandboxInstall, when using the overlayfs backend, can create temporary
extended attributes marking certain files or directories opaque (that
is, they are hidden from the union mount as if they have been deleted).
Unless we reset those xattrs from the target installation directory, we
may have problems with Runner when it creates a union sandbox using the
sandbox overlay.

To be on the safe side we also use 'cp --no-preserve=xattr'. Note that
'mv' is still used elsewhere, so the loop that removes the xattrs may be
still needed anyway.

Fixes #43.