Compile: add support for 'pure' builds
This commit allows one to construct a /System/Index tree based on the
dependencies listed on (1) the given recipe's Dependencies and
BuildDependencies files, (2)
/Programs/Compile/Current/Resources/Dependencies, and (3)
/Programs/Scripts/Current/Resources/Dependencies. The list of
dependencies is processed by `CheckDependencies`, saved to a temporary
file and passed to `Runner` so it builds its sandbox using that set of
programs.

Note that this is still an experimental feature and that bugs are likely
to appear sooner or later. There are at least two outstanding issues
with our integration with `CheckDependencies` that are yet to be fixed,
for instance.