Compile: remove temporary file used by CheckDependencies
The previous commit introduced a temporary file that was not properly
removed by any of the wrap_* functions. The fix was to export the path
to that temporary file so that the Compile instance spawned with Runner
deletes it: spawn_with_runner() itself does some trickery to prevent
running the same hooks twice.