Choose best version number using `GuessLatest`
Use `GuessLatest` rather than `ls --sort=version` to choose the best version number (if none is provided). The advantage is that `ls --sort=version` prefers developement versions like git, svn, cvs, and so on. But in most cases we want the latest stable version, not some obscure developement version. `GuessLatest` will also ensure we are more flexible in the future, should more sophisicated sorting be necessairy or desired