BuildType_meson: Fix potentially destructive sed operation
A sed operation was set to replace all entries containing $target with $goboIndex, however that was potentially harmful! If the user set any value with $settings_target for instance, then part of the path will be substituted with /usr, eg /Programs/Fontconfig/2.15.0/Resources/Default/Settings becomes /usr/Resources/Default/Settings which is an invalid path!

Therefore, by default, set --prefix=/usr in meson options but let the recipe author override it if necessairy.